copied all files to folders
