--- navbargl/src/App.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import './App.css';
-import Navbar from './Navbar'
-function App() {
-  return (
-    <Navbar />
-  );
-}
-
-export default App;
