--- navbargl/src/Navbar.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import './App.css';//put some styling in APP.css
-//this created the hyperlinks.
-function Navbar(){
-    return(
-
-    <div className= "Hello">
-        
-        
-        
-        <a href="#news">App</a>  
-        <a href="#contact">Demo</a>
-        <a href="#contact">Contact</a>
-
-    
-
-    </div>
-
-    );
-
-
-
-
-}
-
-export default Navbar;
